ocr-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
easyocr==1.7.1
opencv-python-headless==4.9.0.80
Pillow==10.3.0
openpyxl==3.1.2
pyyaml==6.0.1
pydantic==2.7.1
rapidfuzz==3.9.3
python-multipart==0.0.9
aiofiles==23.2.1
